o Added support for weeding out potential artifacts that don't have an extension matching the '...s' pattern in the legacy layout (eg. things in '/jars/' that don't have a .jar extension).
o Added artifacts to my local testing repository (which I will be sync'ing with the one on beaver) to satisfy the dotted groupId, the no-extension test, the '.txt' extension test.

o not sure what to do with the last test-case in MNG-312, though, given that I thought we decided not to convert distributions, and the classifiers currently only happen in distros. Will take this discussion to jira, though.

PR: MNG-312


git-svn-id: http://svn.apache.org/repos/asf/maven/components/trunk@163988 13f79535-47bb-0310-9956-ffa450edef68