Resolving: MNG-900
Applied Hiram's patch, with small formatting changes, and a relocation of the logic setting the artifact repository to null (it was being done on the condition that getLogger().isDebugEnabled() == true, just a minor oversight).

There may be a deeper issue here with missing metadata on all repositories...I've annotated what I believe is the root causebut I'm not sure how to fix it, or that it should be fixed.



git-svn-id: http://svn.apache.org/repos/asf/maven/components/trunk@290815 13f79535-47bb-0310-9956-ffa450edef68