PR: MNG-1466 Submitted By: Garrett Conaty Reviewed By: John Casey
Applied patch, with small changes to the default repository id's to match those in Maven proper. "default-local" -> "local", and "default-remote" -> "central".

This patch adds ID handling to repositories in the maven ant tasks. It should enable definition of multiple remote repositories in an Ant script.

Thanks, Garrett and Konstantin, for the work!


git-svn-id: http://svn.apache.org/repos/asf/maven/components/trunk@345281 13f79535-47bb-0310-9956-ffa450edef68