PR: MNG-1311
Added a stack of lifecycle fork-points that will be removed from each successively deeper level of forking in a cumulative manner, to prevent multi-node cycles in addition to simply blocking two-node cycles.

The basic problem with the recursion check that was in there was a type mismatch on the contents of a java.util.List. It's fixed now.


git-svn-id: http://svn.apache.org/repos/asf/maven/components/trunk@355383 13f79535-47bb-0310-9956-ffa450edef68