Move osgi bundle manifest generation to osgi profile.
The generated manifests are incorrect, for instance, maven-project's manifest
misses the package-import org.apache.maven.artifact.transform which should come from 
maven-artifact CacheAware*.java.

Further, this doesn't work in an osgi environment, since if each artifact
is a bundle, then maven-embedder must have access to all component.xml files
in all maven core artifacts, and in all dependencies (which aren't bundles
but that can be fixed).
Osgi import/exports for packages work for resources too, but META-INF/ isn't
a package so this doesn't work. This may work with 'Require-Bundle'
(it does in my implementation although i don't think i follow the spec entirely ;)).
The next problem you get is classcasts for wagon (httplightweightwagon). This
has to be included in a bundle too and get proper exports. The same goes
for all other dependencies (plexus-*).
Maven and plexus are not suited for osgi yet - too much changing is required
in both the metadata structure, the code, and all dependencies and/or
packaging and binary inclusion of dependencies to make this work.
The best bet is to use the one-artifact embedder.



git-svn-id: http://svn.apache.org/repos/asf/maven/components/trunk@549982 13f79535-47bb-0310-9956-ffa450edef68