Add get/setParentContainer to the configuration, and use it in the embedder.
If the classrealm that's passed in has a parent realm, 
then only the resources from the realm are scanned (getRealmResources).
This was done to re-use components from parent realms.

However, when embedding maven within another plexus container,
providing a new dummy classworld with the current classloader
as the parent, no components will be found. Since there's no 
parent container, it won't be checked for components.

So, allowing a parent realm to be set in classworlds also
requires a parent container to be set, so that when no components
are found (since they're present in the parent container),
there is a parent container available to delegate lookups to.


git-svn-id: http://svn.apache.org/repos/asf/maven/components/trunk@563173 13f79535-47bb-0310-9956-ffa450edef68